Move pagination to infraUI/display
