+++ assets/src/infrastructure/ui/display/pagination/types.ts
