--- assets/src/infrastructure/ui/navigation/index.ts
@@ -1,1 +1,0 @@
-export * from './pagination';
